Make factory class to create Person object from post request
Abstracts the creation of a Person object from the resource package (i.e., the PersonResource class), by creating a new class that functions as the factory of Person objects. The factory class uses marhsmallow's deserializiation to map the request json directly into a newly created Person object. This permits avoiding creation of dictionary when taking a post request, to map it to the new Person object. Thus it allows regaining control, compared to the previous design of the program.

Acknowledgements: Marcelo Inhaquites helped all the way.