Add person builder with validation, between resource and model
This implements a builder class (whose name is up for discussion) for Person. This class does validation for request fields and specifies their respective attribute names. Code for actually assigning values to Person attributes remains the same (i.e., very concise).

Note that a class for validating Pet-related fields is also implemented.

It remains to be discussed (1) how Pet object will be created when assigning values to Person attributes; (2) whether splitting the PersonBuilder class into a validator and a person maker would be ideal.