Specify paths for APIspec from blueprint, not hardcoded
Add path to APIspec based on blueprint variables, rather than hardcoding it. This is now working even for the endpoint that is dynamic (the people/<id> endpoint).

This way, whenever the endpoints are edited on blueprints, they are automatically edited on the main __init__ of the app (being able to be edited on a single place).