Move put method to PersonResource and obtain id from path
Move put method from PeopleResource to PersonResource, which is associated with the /people/<id> endpoint.

Id is obtained from path, so it is no longer passed in the body. Therefore it is no longer a field in the UpdatePersonSchema class.

As a result, the put request's body looks exactly like the post request's body.